+++ Root/src/raw_threshold_scan.cpp
@@ -7,7 +7,7 @@
 #include <iomanip>
 
 #include "anyoption.h"
-#include "event.h"
+#include "ocaEvent.h"
 
 AnyOption *opt; // Handle the input options
