+++ src/calibration.cpp
@@ -16,7 +16,7 @@
 #include "TKey.h"
 #include "TPaveText.h"
 #include "anyoption.h"
-#include "event.h"
+#include "ocaEvent.h"
 
 AnyOption *opt; // Handle the option input
