+++ src/raw_clusterize.cpp
@@ -16,7 +16,7 @@
 #include "TTreeReader.h"
 
 #include "anyoption.h"
-#include "event.h"
+#include "ocaEvent.h"
 
 AnyOption *opt; // Handle the input options
