+++ src/raw_cn.cpp
@@ -8,7 +8,7 @@
 #include <iostream>
 
 #include "anyoption.h"
-#include "event.h"
+#include "ocaEvent.h"
 
 AnyOption *opt; //Handle the option input
